Font files can now be read. Each glyph can be rendered to a file.
Next steps are to have a dictionary to store them all in as well as being able to pick and choose in advance which characters (by extension, need to be able to identify each character).
From there, render to screen. This will require storing it all in a structure that takes into account the glyph properties, like where it sits on the line. As well as this, being able to choose a font size and then scaling it down as and when (or perhaps creating it on the fly? Might be slow.)